+++ Tests/SymfonyApp/TestBundle/Hautelook/AliceBundle/Tests/SymfonyApp/TestBundle/Document/Product.php
@@ -1,0 +1,81 @@
+<?php
+
+namespace Hautelook\AliceBundle\Tests\SymfonyApp\TestBundle\Document;
+
+
+
+/**
+ * Hautelook\AliceBundle\Tests\SymfonyApp\TestBundle\Document\Product
+ */
+class Product
+{
+    /**
+     * @var $id
+     */
+    protected $id;
+
+    /**
+     * @var string $name
+     */
+    protected $name;
+
+    /**
+     * @var float $price
+     */
+    protected $price;
+
+
+    /**
+     * Get id
+     *
+     * @return id $id
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * Set name
+     *
+     * @param string $name
+     * @return self
+     */
+    public function setName($name)
+    {
+        $this->name = $name;
+        return $this;
+    }
+
+    /**
+     * Get name
+     *
+     * @return string $name
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * Set price
+     *
+     * @param float $price
+     * @return self
+     */
+    public function setPrice($price)
+    {
+        $this->price = $price;
+        return $this;
+    }
+
+    /**
+     * Get price
+     *
+     * @return float $price
+     */
+    public function getPrice()
+    {
+        return $this->price;
+    }
+}