+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 * Upgraded [nelmio/alice](https://github.com/nelmio/alice) to the version [2.0.0](https://github.com/nelmio/alice/releases/tag/2.0.0)
 * Now uses [`Nelmio\Alice\Fixtures`](https://github.com/nelmio/alice/blob/master/src/Nelmio/Alice/Fixtures.php) and
 its loaders instead of custom ones
-* Introduced a command `hautelook_alice:fixtures:load`
+* Add support for [Doctrine ODM (MongoDB)](http://doctrine-mongodb-odm.readthedocs.org/en/latest/)
+* Add support for [Doctrine PHPCR ODM](http://doctrine-phpcr-odm.readthedocs.org/en/latest/)
+* Introduced commands:
+	* `hautelook_alice:doctrine:fixtures:load`
+	* `hautelook_alice:doctrine:mongodb:fixtures:load`
+	* `hautelook_alice:doctrine:phpcr:fixtures:load`
 * Possibility to load fixtures by environment
 * Possibility to load fixtures by bundle
 * No longer need to create a data loader to load fixtures